pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pyveb"
version = "0.1.26"
authors = [
  { name="pieter de petter", email="pieter.de.petter@veb.be" },
]
description = "Package containing common code and reusable components for pipelines and dags"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
   "selenium==3.141.0", 
   "webdriver-manager==3.4.1", 
   "boto3==1.24.24", 
   "pyyaml==6.0",
   "simple-ddl-parser==0.26.5", 
   "pyspark",
   "psycopg2-binary==2.9.1",
   "psutil",
   "pandas==1.3.2",
   "numpy==1.21.2",
   "s3fs",
   "pyarrow==5.0.0",
   "requests==2.27.1",
   "pyodbc==4.0.30"
]

[project.urls]
"Homepage" = "https://vlaamsenergiebedrijf.visualstudio.com/Terra/_git/terra-etl?path=/common_code"
